- Swap argument order of connect_alsa. - Use local environment script if file available. This decouple the local config from the files in this repository.

+++ scripts/environment.sh
@@ -4,28 +4,36 @@
 # GridMusic assumes that it is installed under
 # ${KEYROOT}/contrib/GridMusic
 
-#export KEYROOT=/opt/kinect/keykit/
-export KEYROOT=$(pwd)/../../..
-export KEYROOT=$HOME/software/keykit
+# Use local environment file if available. Use this file as template, but do not add the local file to git repository.
+LOCAL_FILE="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )""/environment.local.sh"
+if [ -e "${LOCAL_FILE}" -a "$AVOID_INF_LOOP" != "1" ] ; then
+  AVOID_INF_LOOP=1
+  source "${LOCAL_FILE}"
+  AVOID_INF_LOOP=0
+else
 
-# Linux
-export KEYKIT=key_alsa
-export LOWKEYKIT=${KEYKIT}
+  export KEYROOT=$(pwd)/../../..
 
-# Windows
-#export KEYKIT=key.exe
-#export LOWKEYKIT=lowkey.exe
+  # Linux
+  export KEYKIT=key_alsa
+  export LOWKEYKIT=${KEYKIT}
 
-# Set KEYPATH to parse GridMusic files at startup.
-# Alternatively, you can place '#include contrib/GridMusic/start.k'
-# into the keyrc.k file, see below.
-DEFAULT_KEYPATH=".:${KEYROOT}/lib:${KEYROOT}/liblocal" # could be differ on Windows?!
-THIS_DIR=$(pwd)
-export KEYPATH="${DEFAULT_KEYPATH}:${THIS_DIR%/scripts}:${KEYPATH}"
+  # Windows
+  #export KEYKIT=key.exe
+  #export LOWKEYKIT=lowkey.exe
 
-# Copy lib/keyrc.k into KEYROOT if file not exists
-if [ ! -e ${KEYROOT}/keyrc.k ] ; then
-	echo "keyrc.k not found in ${KEYROOT}! Copy default keyrc.k from lib subdirectory..."
-	cp ${KEYROOT}/lib/keyrc.k ${KEYROOT}
+
+  # Set KEYPATH to parse Pyconsole files at startup.
+  # Alternatively, you can place '#include contrib/Pyconsole/start.k'
+  # into the keyrc.k file, see below.
+  DEFAULT_KEYPATH=".:${KEYROOT}/lib:${KEYROOT}/liblocal" # could be differ on Windows?!
+  THIS_DIR=$(pwd)
+  export KEYPATH="${DEFAULT_KEYPATH}:${THIS_DIR%/scripts}:${KEYPATH}"
+
+  # Copy lib/keyrc.k into KEYROOT if file not exists
+  if [ ! -e ${KEYROOT}/keyrc.k ] ; then
+    echo "keyrc.k not found in ${KEYROOT}! Copy default keyrc.k from lib subdirectory..."
+    cp ${KEYROOT}/lib/keyrc.k ${KEYROOT}
+  fi
+
 fi
-
